more aggressive typechecks (#44)
Fixes #42.

I'm basing this on the es2016 branch for convenience, even though don't I really intend to make a release of that line before bumping to es2017. This is a breaking change, since the whole point is to reject inputs which it previously accepted.

I recommend reading the generated file to get a sense of what the typechecks look like.